safari-600.7-branch/LayoutTests/platform/ios-sim/storage/indexeddb/database-wrapper-expected.txt

CONSOLE MESSAGE: line 14: TypeError: undefined is not an object (evaluating 'deleteRequest.onblocked = unexpectedBlockedCallback')
Ensure IDBDatabase wrapper isn't prematurely collected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


dbname = "database-wrapper.html"
indexedDB.deleteDatabase(dbname)
FAIL indexedDB.deleteDatabase(dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'deleteRequest.onblocked = unexpectedBlockedCallback')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
